ml-backend/requirements.txt
Flask==2.3.2
tensorflow==2.15.0
pillow
numpy
opencv-python
pymongo
flask-cors
